scripts/palette_extract/requirements.txt
numpy>=1.24,<3
Pillow>=10,<12
PyYAML>=6,<7
